+++ libs/angular-components/tedi/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./base";
+export * from "./content";
+export * from "./layout";
+export * from "./loader";
+export * from "./navigation";